backend/requirements.txt
django==3.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-decouple
cloudinary
